requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curlify==2.2.1
exceptiongroup==1.2.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pytest==7.4.3
python-dotenv==1.0.0
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
